requirements.txt
asgiref==3.6.0
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.6
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.2
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-writable-nested==0.7.0
execnet==1.9.0
Faker==12.0.1
gunicorn==20.1.0
hypothesis==6.75.5
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
mixer==7.2.2
model-bakery==1.11.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyscopg2==66.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
sqlparse==0.4.4
stripe==5.4.0
termcolor==2.3.0
toml==0.10.2
typing_extensions==4.5.0
upgrade-requirements==1.7.0
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.4.0
